requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
dataclasses-json==0.5.8
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.98.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.202
langchainplus-sdk==0.0.10
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
orjson==3.8.9
packaging==23.1
passlib==1.7.4
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg-pool==3.1.7
psycopg2-binary==2.9.6
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.27.0
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
ujson==5.7.0
urllib3==2.0.3
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0
yarl==1.9.2
